Title: Adding solution for hour glass challegen From: HackerRank -https://www.hackerrank.com/challenges/2d-array/problem
Description:

Given a  2D Array, :

1 1 1 0 0 0
0 1 0 0 0 0
1 1 1 0 0 0
0 0 0 0 0 0
0 0 0 0 0 0
0 0 0 0 0 0

We define an hourglass in  to be a subset of values with indices falling
in this pattern in 's graphical representation:

a b c
  d
e f g

There are  hourglasses in , and an hourglass sum is the sum of an
hourglass' values. Calculate the hourglass sum for every hourglass in ,
then print the maximum hourglass sum.